+++ lib/pack_app/screens/app_loading/app_loading_screen.dart
@@ -6,26 +6,23 @@ class AppLoadingScreen extends StatelessWidget {
   }) : super(key: key);
   @override
   Widget build(final BuildContext context) {
-    final settings = SettingsStateScope.of(context);
+    final settings = context.watch<GeneralSettingsController>();
     final statusText = appLoadingStatusesTitles[settings.loadingStatus]
             ?.getByLanguage(intl.Intl.systemLocale) ??
         '';
 
-    return Container(
+    return ColoredBox(
       color: AppColors.black,
       child: Center(
-        child: Directionality(
-          textDirection: TextDirection.ltr,
-          child: Column(
-            mainAxisSize: MainAxisSize.min,
-            children: [
-              const CircularProgressIndicator(
-                valueColor: AlwaysStoppedAnimation(AppColors.primary2),
-              ),
-              const SizedBox(height: 5),
-              Text(statusText),
-            ],
-          ),
+        child: Column(
+          mainAxisSize: MainAxisSize.min,
+          children: [
+            const CircularProgressIndicator(
+              valueColor: AlwaysStoppedAnimation(AppColors.primary2),
+            ),
+            const SizedBox(height: 5),
+            Text(statusText),
+          ],
         ),
       ),
     );