+++ lib/pack_idea/widgets/question_dropdown.dart
@@ -13,15 +13,20 @@ class _QuestionDropdown extends HookWidget {
   @override
   Widget build(final BuildContext context) {
     final chosenQuestion = useState(answer.question);
+
     useEffect(
       () {
         chosenQuestion.value = answer.question;
+
+        return null;
       },
       [answer.question],
     );
+
     final ideaQuestionsProvider = context.read<IdeaProjectQuestionsProvider>();
     final questions = ideaQuestionsProvider.values;
     final textStyle = Theme.of(context).textTheme.bodyText1!;
+
     final questionsItems = questions.map(
       (final question) => DropdownMenuItem<IdeaProjectQuestion>(
         value: question,