+++ lib/pack_note/pack_note.dart
@@ -3,6 +3,7 @@ library pack_note;
 import 'dart:async';
 
 import 'package:flutter/cupertino.dart';
+import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_hooks/flutter_hooks.dart';
 import 'package:lastanswer/abstract/abstract.dart';
@@ -17,10 +18,12 @@ import 'package:lastanswer/utils/utils.dart';
 import 'package:provider/provider.dart';
 import 'package:recase/recase.dart';
 import 'package:rxdart/rxdart.dart';
+import 'package:universal_io/io.dart';
 
 part 'screens/note_project_screen.dart';
 part 'screens/note_project_screen_state.dart';
 part 'states/use_note_project_updater.dart';
 part 'widgets/desktop_note_settings.dart';
 part 'widgets/mobile_note_settings.dart';
+part 'widgets/note_project_side_actionbar.dart';
 part 'widgets/note_settings_button.dart';