+++ lib/pack_note/screens/note_project_screen.dart
@@ -40,12 +40,20 @@ class NoteProjectScreen extends HookWidget {
       checkIsProjectActive: checkIsProjectActive,
     );
 
+    double? appBarHeight;
+    if (Platform.isMacOS) {
+      appBarHeight = null;
+    } else {
+      appBarHeight = screenLayout.small ? 40 : 30;
+    }
+
     return Scaffold(
       backgroundColor: theme.canvasColor,
       restorationId: 'notes/$noteId',
+      resizeToAvoidBottomInset: false,
       appBar: BackTextUniversalAppBar(
         useBackButton: true,
-        height: screenLayout.small ? null : 30,
+        height: appBarHeight,
         screenLayout: screenLayout,
         titleStr: '',
         onBack: state.onBack,
@@ -58,70 +66,54 @@ class NoteProjectScreen extends HookWidget {
           child: SpecialEmojisKeyboardActions(
             focusNode: noteFocusNode,
             controller: noteController,
-            child: Column(
-              mainAxisAlignment: MainAxisAlignment.end,
-              crossAxisAlignment: CrossAxisAlignment.stretch,
-              children: [
-                Expanded(
-                  child: Row(
-                    crossAxisAlignment: CrossAxisAlignment.end,
-                    mainAxisAlignment: MainAxisAlignment.center,
-                    children: [
-                      const SizedBox(width: 20),
-                      Expanded(
-                        child: ProjectTextField(
-                          hintText: S.current.writeANote,
-                          fillColor: Colors.transparent,
-                          filled: false,
-                          limit: note.value.charactersLimit,
-                          focusNode: noteFocusNode,
-                          endlessLines: true,
-                          onSubmit: state.onBack,
-                          controller: noteController,
+            builder: (
+              final context,
+              final showEmojiKeyboard,
+              final closeEmojiKeyboard,
+              final isEmojiKeyboardOpen,
+            ) {
+              return Column(
+                mainAxisAlignment: MainAxisAlignment.end,
+                crossAxisAlignment: CrossAxisAlignment.stretch,
+                children: [
+                  Expanded(
+                    child: Row(
+                      crossAxisAlignment: CrossAxisAlignment.end,
+                      mainAxisAlignment: MainAxisAlignment.center,
+                      children: [
+                        const SizedBox(width: 20),
+                        Expanded(
+                          child: ProjectTextField(
+                            hintText: S.current.writeANote,
+                            fillColor: Colors.transparent,
+                            filled: false,
+                            limit: note.value.charactersLimit,
+                            focusNode: noteFocusNode,
+                            endlessLines: true,
+                            focusOnInit: note.value.note.isEmpty,
+                            onSubmit: state.onBack,
+                            controller: noteController,
+                          ),
                         ),
-                      ),
-                      Column(
-                        mainAxisAlignment: MainAxisAlignment.end,
-                        children: [
-                          NoteSettingsButton(
-                            note: note.value,
-                            onRemove: state.onRemove,
-                            updatesStream: updatesStream,
-                          ),
-                          SpecialEmojiPopup(
-                            controller: noteController,
-                            focusNode: noteFocusNode,
-                          ),
-                          SizedBox(
-                            height: 34,
-                            width: 48,
-                            child: IconShareButton(
-                              onTap: () {
-                                ProjectSharer.of(context)
-                                    .share(project: note.value);
-                              },
-                            ),
-                          ),
-                          EmojiPopup(
-                            controller: noteController,
-                            focusNode: noteFocusNode,
-                          ),
-                        ]
-                            .map(
-                              (final e) => Padding(
-                                padding: const EdgeInsets.only(top: 15),
-                                child: e,
-                              ),
-                            )
-                            .toList(),
-                      ),
-                    ],
+                        NoteProjectSideActionBar(
+                          closeEmojiKeyboard: closeEmojiKeyboard,
+                          isEmojiKeyboardOpen: isEmojiKeyboardOpen,
+                          note: note,
+                          noteController: noteController,
+                          noteFocusNode: noteFocusNode,
+                          onRemove: state.onRemove,
+                          showEmojiKeyboard: showEmojiKeyboard,
+                          updatesStream: updatesStream,
+                        ),
+                      ],
+                    ),
                   ),
-                ),
-                const SizedBox(height: 14),
-                const BottomSafeArea(),
-              ],
-            ),
+                  if ((isNativeDesktop || kIsWeb) && !isEmojiKeyboardOpen.value)
+                    const SizedBox(height: 16),
+                  if (!isEmojiKeyboardOpen.value) const BottomSafeArea(),
+                ],
+              );
+            },
           ),
         ),
       ),